Add NoZygote command line option
Quoting from: https://codereview.chromium.org/2384163002

This patch adds a --no-zygote command line flag which can be used to
completely disable the use of the zygote process. This is needed for
running Headless Chrome in environments where maintaining a forking
server such as the zygote is unnecessary and/or impractical.